requirements.txt
autopep8 
pylint 
black
flake8
bandit
fastapi
uvicorn
google-generativeai
python-dotenv
numpy
scikit-learn
pymongo
pybars3
sentence_transformers